sftp: get windows compiling and allow tests to run that can pass
Separate out windows specific calls to separate build files.
Unix specific calls, such as unix sockets or unix syscalls
give windows equivalents. Do not yet provide a windows
equivalent to createRequestServerPair as doing so exposes
other windows breakages. For now simply get it to compile on
windows. Passing the client-server tests will be handled in a follow
up PR.

Fixes #157
